Transitioning destroy to the Factory / Client level
