+++ client/src/main/java/io/split/client/SplitClient.java
@@ -90,4 +90,11 @@ public interface SplitClient {
      * @return the evaluated treatment, the default treatment of this feature, or 'control'.
      */
     String getTreatment(Key key, String split, Map<String, Object> attributes);
+
+
+    /**
+     * Destroys the background processes and clears the cache, releasing the resources used by
+     * the any instances of SplitClient or SplitManager generated by the client's parent SplitFactory
+     */
+    void destroy();
 }
