+++ client/src/main/java/io/split/client/SplitFactory.java
@@ -6,4 +6,5 @@ package io.split.client;
 public interface SplitFactory {
     SplitClient client();
     SplitManager manager();
+    void destroy();
 }
